Update COCO json format
I converted my own dataset to the COCO format. When I tried to fine-tune with it, I received errors from pycocotools/cocoeval.py. One of the errors was fixed by the key "iscrowd" which should always be 0 because we're using bounding boxes. The other one was the "area" key which should be width * height. Both values are to the best of my knowledge. Maybe pycocotools made breaking changes or I've set some options wrong. In any case, the suggested change worked for me.